+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     - Interval Discretizer variable or equi-depth binning
 
 - 1.1.2
+    - Improved random floating point number precision
     - Deduplicate Preset seeder centroids
     - Fix Gradient Boost learning rate upper bound
     - Fix Loda histogram edge alignment